Make fixSystemBarsPadding not accept a null View
Now that we use BaseFragment which uses this never being null, only a few null usages left, which we just make sure aren't null here. Being explicit and not accepting null makes it easier to know what type is being passed and debug any potential issues as well which is why I wrote BaseFragment to not accept nullable for `fixPadding`.