Add a public.xml
This marks all resources as private by default, it won't cause build errors for any extensions.

In the future we can define resources we explicitly want public in order ro create a stable API for resource access. When resources are changed or removed or even reordered in some cases, the internal resource ID will change, which will lead to runtime errors if an extension is compiled with it. Using public.xml would also allow us to define fixed IDs at a later time to prevent the potential for such issues as well.